SDK-737/ Fix folder and file naming in GO
