+++ .github/workflows/ci.yml
@@ -45,7 +45,7 @@ jobs:
       run: |
         cd commonutils
         go build -v ./...
-        cd ../service-account
+        cd ../serviceaccount
         go build -v ./...
         cd ../skyflow
         go build -v ./...
@@ -53,7 +53,7 @@ jobs:
 
     - name: 'Run Tests'
       run: |
-        cd service-account
+        cd serviceaccount
         go test ./... -coverprofile=coverage.out
         cd ../skyflow
         go test ./... -coverprofile=coverage.out
@@ -62,6 +62,6 @@ jobs:
       uses: codecov/codecov-action@v2.1.0
       with:
         token: ${{ secrets.CODECOV_REPO_UPLOAD_TOKEN }}
-        files: service-account/coverage.out, skyflow/coverage.out
+        files: serviceaccount/coverage.out, skyflow/coverage.out
         name: codecov-skyflow-go
         verbose: true
