+++ samples/go.mod
@@ -4,6 +4,6 @@ go 1.13
 
 require (
 	github.com/skyflowapi/skyflow-go v1.0.0 // indirect
-	github.com/skyflowapi/skyflow-go/service-account v0.0.0-20220315114742-246a206a0e88 // indirect
+	github.com/skyflowapi/skyflow-go/serviceaccount v0.0.0-20220315114742-246a206a0e88 // indirect
 	github.com/skyflowapi/skyflow-go/skyflow v0.0.0-20220315112934-3c8113d9d693 // indirect
 )
