+++ samples/vaultapi/detokenize.go
@@ -1,5 +1,5 @@
 /*
-	Copyright (c) 2022 Skyflow, Inc. 
+Copyright (c) 2022 Skyflow, Inc.
 */
 package main
 
@@ -7,7 +7,7 @@ import (
 	"fmt"
 
 	logger "github.com/skyflowapi/skyflow-go/commonutils/logwrapper"
-	saUtil "github.com/skyflowapi/skyflow-go/service-account/util"
+	saUtil "github.com/skyflowapi/skyflow-go/serviceaccount/util"
 	Skyflow "github.com/skyflowapi/skyflow-go/skyflow/client"
 	"github.com/skyflowapi/skyflow-go/skyflow/common"
 )
@@ -28,29 +28,28 @@ func GetToken() (string, error) {
 	}
 	return bearerToken, nil
 }
-
 func main() {
 
 	defer func() {
-		if err := recover(); err != nil {
-			fmt.Println("error : ", err)
+		if r := recover(); r != nil {
+			fmt.Println("error: ", r)
 		}
 	}()
+
 	logger.SetLogLevel(logger.INFO) //set loglevel to INFO
 	configuration := common.Configuration{VaultID: "<vault_id>", VaultURL: "<vault_url>", TokenProvider: GetToken}
 	var client = Skyflow.Init(configuration)
-	var options = common.InsertOptions{Tokens: false}
 	var records = make(map[string]interface{})
-	var record = make(map[string]interface{})
-	record["table"] = "cards"
-	var fields = make(map[string]interface{})
-	fields["cvv"] = "123"
-	fields["fullname"] = "name"
-	record["fields"] = fields
+	var record1 = make(map[string]interface{})
+	record1["token"] = "<token>"
+	var record2 = make(map[string]interface{})
+	record2["token"] = "<token>"
+
 	var recordsArray []interface{}
-	recordsArray = append(recordsArray, record)
+	recordsArray = append(recordsArray, record1)
+	recordsArray = append(recordsArray, record2)
 	records["records"] = recordsArray
-	res, err := client.Insert(records, options)
+	res, err := client.Detokenize(records)
 	if err == nil {
 		fmt.Println("Records : ", res.Records)
 	} else {
