+++ samples/vaultapi/insert.go
@@ -1,5 +1,5 @@
 /*
-	Copyright (c) 2022 Skyflow, Inc. 
+Copyright (c) 2022 Skyflow, Inc.
 */
 package main
 
@@ -7,7 +7,7 @@ import (
 	"fmt"
 
 	logger "github.com/skyflowapi/skyflow-go/commonutils/logwrapper"
-	saUtil "github.com/skyflowapi/skyflow-go/service-account/util"
+	saUtil "github.com/skyflowapi/skyflow-go/serviceaccount/util"
 	Skyflow "github.com/skyflowapi/skyflow-go/skyflow/client"
 	"github.com/skyflowapi/skyflow-go/skyflow/common"
 )
@@ -32,23 +32,25 @@ func GetToken() (string, error) {
 func main() {
 
 	defer func() {
-		if r := recover(); r != nil {
-			fmt.Println("error: ", r)
+		if err := recover(); err != nil {
+			fmt.Println("error : ", err)
 		}
 	}()
 	logger.SetLogLevel(logger.INFO) //set loglevel to INFO
 	configuration := common.Configuration{VaultID: "<vault_id>", VaultURL: "<vault_url>", TokenProvider: GetToken}
 	var client = Skyflow.Init(configuration)
+	var options = common.InsertOptions{Tokens: false}
 	var records = make(map[string]interface{})
-	var record1 = make(map[string]interface{})
-	record1["ids"] = []interface{}{"<id1>", "<id2>"}
-	record1["table"] = "cards"
-	record1["redaction"] = "PLAIN_TEXT"
-
+	var record = make(map[string]interface{})
+	record["table"] = "cards"
+	var fields = make(map[string]interface{})
+	fields["cvv"] = "123"
+	fields["fullname"] = "name"
+	record["fields"] = fields
 	var recordsArray []interface{}
-	recordsArray = append(recordsArray, record1)
+	recordsArray = append(recordsArray, record)
 	records["records"] = recordsArray
-	res, err := client.GetById(records)
+	res, err := client.Insert(records, options)
 	if err == nil {
 		fmt.Println("Records : ", res.Records)
 	} else {
