+++ skyflow/client/client_test.go
@@ -1,5 +1,5 @@
 /*
-	Copyright (c) 2022 Skyflow, Inc. 
+Copyright (c) 2022 Skyflow, Inc.
 */
 package client
 
@@ -12,7 +12,7 @@ import (
 	"github.com/joho/godotenv"
 	errors1 "github.com/skyflowapi/skyflow-go/commonutils/errors"
 	"github.com/skyflowapi/skyflow-go/commonutils/messages"
-	saUtil "github.com/skyflowapi/skyflow-go/service-account/util"
+	saUtil "github.com/skyflowapi/skyflow-go/serviceaccount/util"
 	"github.com/skyflowapi/skyflow-go/skyflow/common"
 )
